+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - slightly rework file dialog GUI
+- core/DebugScene : make setupPipelines() private, do not use optional
 
 ### Fixed